Feature/deleteuser2

+++ client/src/containers/JournalContainer.js
@@ -47,23 +47,32 @@ const JournalContainer = () => {
       console.error("Error posting new user:", error);
     }
   };
+  
+  //second attempt at deleteAccount
 
-  const handleDeleteAccount = async (user) => {
+  const handleDeleteAccount = async (id) => {
+    if (currentUser === null) {
+      alert('Not signed in!');
+      return;
+    }
+  
+    // Get the current user's ID
+    const userId = currentUser.id;
+  
+    // Send a DELETE request to the /users/{id} endpoint
     try {
-      const response = await fetch(`http://localhost:8080/users/${currentUser.id}`, {
-        method: "DELETE",
-        headers: { "Content-Type": "application/json" },
+      const response = await fetch(`http://localhost:8080/users/${id}`, {
+        method: 'DELETE',
       });
-
+  
       if (response.status === 204) {
-        alert("Account deleted successfully!");
-        setCurrentUser(null);
-        setJournalEntries([]);
+        alert('Account deleted successfully!');
+        currentUser = null;
       } else {
-        console.error("Failed to delete account. Status code:", response.status);
+        console.error('Failed to delete account. Status code:', response.status);
       }
     } catch (error) {
-      console.error("Error deleting account:", error);
+      console.error('Failed to delete account:', error);
     }
   };
 
@@ -156,9 +165,9 @@ const JournalContainer = () => {
       }
     );
 
-    const entryIndex = journalEntries.indexOf(entry);
-    const updatedJournalEntries = journalEntries;
-    updatedJournalEntries.splice(entryIndex, 1, entry);
+    const entryIndex = journalEntries.indexOf(entry); // find the index of the entry to be replaced
+    const updatedJournalEntries = journalEntries;     // make a copy
+    updatedJournalEntries.splice(entryIndex, 1, entry); //We remove one entry at the index and replace it with entry
     setJournalEntries(updatedJournalEntries);
   };
 
@@ -172,11 +181,6 @@ const JournalContainer = () => {
     );
     setJournalEntries(journalEntries.filter((entry) => entry.id !== entryId));
   };
-
-  // useEffect(() => {
-  //   fetchUserById(2); //Remember to take this out after
-  //   fetchAllEntriesByUserId(2);
-  // }, []);
 
   const journalEntryRoutes = createBrowserRouter([
     {
@@ -204,9 +208,7 @@ const JournalContainer = () => {
 
         {
           path: "/entries/new",
-          element: (
-            <NewEntryForm submitForm={postNewEntry} currentUser={currentUser} />
-          ),
+          element: <NewEntryForm submitForm={postNewEntry} currentUser={currentUser} />,
         },
         {
           path: "/entries/:id/edit",
